// sdk/node/src/generated/protobuf/transaction/v1/transaction.ts
export interface Transaction {
  $type: "protobuf.transaction.v1.Transaction";
  device:
    | Device
    | undefined;
  /** Required transaction core (id, amount, datetime, channel, type). */
  transaction: TransactionData | undefined;
  creditCard: CreditCard | undefined;
  boleto: Boleto | undefined;
  order: Order | undefined;
  pos: Pos | undefined;
  customer: Customer | undefined;
  extraData: { [key: string]: any } | undefined;
}

export interface Device {
  $type: "protobuf.transaction.v1.Device";
  fingerprint: string;
  vendor: string;
  browser: string;
  category: string;
  ip: string;
  imei: string;
}

export interface TransactionData {
  $type: "protobuf.transaction.v1.TransactionData";
  id: string;
  description: string;
  currencyCode: string;
  /** Money as a positive decimal string (source validated amount_total > 0). */
  amountTotal: string;
  instalments: number;
  datetimeRequest: Date | undefined;
  datetimeProgram: Date | undefined;
  acquirer: string;
  channel: string;
  type: string;
  origin: string;
  /** [latitude, longitude] */
  location: number[];
  payer: Party | undefined;
  receiver: Party | undefined;
}

export interface Party {
  $type: "protobuf.transaction.v1.Party";
  id: string;
  registrationData: RegistrationData | undefined;
  bankingData: BankingData | undefined;
  pixKey: string[];
  pixKeyType: PixKeyType[];
}

export interface CreditCard {
  $type: "protobuf.transaction.v1.CreditCard";
  id: string;
  lastFour: string;
  holderName: string;
  scheme: string;
  country: string;
  bin: string;
  /** MM/YY or MM/YYYY in source; kept as string */
  expirationDate: string;
}

export interface Boleto {
  $type: "protobuf.transaction.v1.Boleto";
  ourNumber: string;
  barCode: string;
  expirationDate: string;
  issueDate: string;
  beneficiary: string;
  payer: string;
}

export interface Order {
  $type: "protobuf.transaction.v1.Order";
  items: OrderItem[];
  delivery:
    | Delivery
    | undefined;
  /** money (decimal string) */
  voucherValue: string;
  multiplePaymentMethods: boolean;
}
// ...
export interface Pos {
  $type: "protobuf.transaction.v1.Pos";
  id: string;
  name: string;
  address: Address | undefined;
}

export interface Customer {
  $type: "protobuf.transaction.v1.Customer";
  name: string;
  email: string;
  datetimeAccountCreate: Date | undefined;
  phoneNumber: string;
  birthdate: string;
  docType: string;
  doc: string;
  id: string;
  address: Address | undefined;
}
// ...
export const Device: MessageFns<Device, "protobuf.transaction.v1.Device"> = {
  $type: "protobuf.transaction.v1.Device" as const,

  encode(message: Device, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.fingerprint !== "") {
      writer.uint32(10).string(message.fingerprint);
    }
    if (message.vendor !== "") {
      writer.uint32(18).string(message.vendor);
    }
    if (message.browser !== "") {
      writer.uint32(26).string(message.browser);
    }
    if (message.category !== "") {
      writer.uint32(34).string(message.category);
    }
    if (message.ip !== "") {
      writer.uint32(42).string(message.ip);
    }
    if (message.imei !== "") {
      writer.uint32(50).string(message.imei);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Device {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDevice();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.fingerprint = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.vendor = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.browser = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.category = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.ip = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.imei = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Device {
    return {
      $type: Device.$type,
      fingerprint: isSet(object.fingerprint) ? globalThis.String(object.fingerprint) : "",
      vendor: isSet(object.vendor) ? globalThis.String(object.vendor) : "",
      browser: isSet(object.browser) ? globalThis.String(object.browser) : "",
      category: isSet(object.category) ? globalThis.String(object.category) : "",
      ip: isSet(object.ip) ? globalThis.String(object.ip) : "",
      imei: isSet(object.imei) ? globalThis.String(object.imei) : "",
    };
  },

  toJSON(message: Device): unknown {
    const obj: any = {};
    if (message.fingerprint !== "") {
      obj.fingerprint = message.fingerprint;
    }
    if (message.vendor !== "") {
      obj.vendor = message.vendor;
    }
    if (message.browser !== "") {
      obj.browser = message.browser;
    }
    if (message.category !== "") {
      obj.category = message.category;
    }
    if (message.ip !== "") {
      obj.ip = message.ip;
    }
    if (message.imei !== "") {
      obj.imei = message.imei;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Device>, I>>(base?: I): Device {
    return Device.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Device>, I>>(object: I): Device {
    const message = createBaseDevice();
    message.fingerprint = object.fingerprint ?? "";
    message.vendor = object.vendor ?? "";
    message.browser = object.browser ?? "";
    message.category = object.category ?? "";
    message.ip = object.ip ?? "";
    message.imei = object.imei ?? "";
    return message;
  },
};
// ...
export const CreditCard: MessageFns<CreditCard, "protobuf.transaction.v1.CreditCard"> = {
  $type: "protobuf.transaction.v1.CreditCard" as const,

  encode(message: CreditCard, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.lastFour !== "") {
      writer.uint32(18).string(message.lastFour);
    }
    if (message.holderName !== "") {
      writer.uint32(26).string(message.holderName);
    }
    if (message.scheme !== "") {
      writer.uint32(34).string(message.scheme);
    }
    if (message.country !== "") {
      writer.uint32(42).string(message.country);
    }
    if (message.bin !== "") {
      writer.uint32(50).string(message.bin);
    }
    if (message.expirationDate !== "") {
      writer.uint32(58).string(message.expirationDate);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreditCard {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreditCard();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.lastFour = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.holderName = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.scheme = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.country = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.bin = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.expirationDate = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreditCard {
    return {
      $type: CreditCard.$type,
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      lastFour: isSet(object.lastFour)
        ? globalThis.String(object.lastFour)
        : isSet(object.last_four)
        ? globalThis.String(object.last_four)
        : "",
      holderName: isSet(object.holderName)
        ? globalThis.String(object.holderName)
        : isSet(object.holder_name)
        ? globalThis.String(object.holder_name)
        : "",
      scheme: isSet(object.scheme) ? globalThis.String(object.scheme) : "",
      country: isSet(object.country) ? globalThis.String(object.country) : "",
      bin: isSet(object.bin) ? globalThis.String(object.bin) : "",
      expirationDate: isSet(object.expirationDate)
        ? globalThis.String(object.expirationDate)
        : isSet(object.expiration_date)
        ? globalThis.String(object.expiration_date)
        : "",
    };
  },

  toJSON(message: CreditCard): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.lastFour !== "") {
      obj.lastFour = message.lastFour;
    }
    if (message.holderName !== "") {
      obj.holderName = message.holderName;
    }
    if (message.scheme !== "") {
      obj.scheme = message.scheme;
    }
    if (message.country !== "") {
      obj.country = message.country;
    }
    if (message.bin !== "") {
      obj.bin = message.bin;
    }
    if (message.expirationDate !== "") {
      obj.expirationDate = message.expirationDate;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreditCard>, I>>(base?: I): CreditCard {
    return CreditCard.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreditCard>, I>>(object: I): CreditCard {
    const message = createBaseCreditCard();
    message.id = object.id ?? "";
    message.lastFour = object.lastFour ?? "";
    message.holderName = object.holderName ?? "";
    message.scheme = object.scheme ?? "";
    message.country = object.country ?? "";
    message.bin = object.bin ?? "";
    message.expirationDate = object.expirationDate ?? "";
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in Exclude<keyof T, "$type">]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P> | "$type">]: never };
// ...
export interface MessageFns<T, V extends string> {
  readonly $type: V;
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}